+++ walkability/core/operator_worker.py
@@ -52,7 +52,8 @@ class OperatorWalkability(BaseOperator[ComputeInputWalkability]):
             aoi=aoi, max_walking_distance=params.max_walking_distance
         )
 
-        try:
+        with self.catch_exceptions(indicator_name='Areal summary charts', resources=resources):
+            areal_summaries = dict()  # Empty dict in case summaries fail
             areal_summaries = summarise_by_area(
                 paths=line_paths,
                 aoi=aoi,
@@ -60,11 +61,6 @@ class OperatorWalkability(BaseOperator[ComputeInputWalkability]):
                 projected_crs=get_utm_zone(aoi),
                 ohsome_client=self.ohsome,
             )
-        except Exception as error:
-            log.error(
-                'The computation of the areal summaries indicator failed. No artifact will be created.', exc_info=error
-            )
-            areal_summaries = dict()
         path_artifacts = build_path_categorisation_artifact(
             paths_line=line_paths,
             paths_polygon=polygon_paths,
@@ -90,20 +86,21 @@ class OperatorWalkability(BaseOperator[ComputeInputWalkability]):
             artifacts.append(detour_artifact)
 
         if WalkabilityIndicators.NATURALNESS in params.indicators_to_compute:
-            naturalness_artifact = naturalness_analysis(
-                line_paths=line_paths,
-                aoi=aoi,
-                index=params.naturalness_index,
-                resources=resources,
-                naturalness_utility=self.naturalness_utility,
-            )
-            artifacts.append(naturalness_artifact)
+            with self.catch_exceptions(indicator_name='Naturalness', resources=resources):
+                naturalness_artifact = naturalness_analysis(
+                    line_paths=line_paths,
+                    index=params.naturalness_index,
+                    resources=resources,
+                    naturalness_utility=self.naturalness_utility,
+                )
+                artifacts.append(naturalness_artifact)
 
         if WalkabilityIndicators.SLOPE in params.indicators_to_compute:
-            slope_artifact = slope_analysis(
-                line_paths=line_paths, aoi=aoi, ors_client=self.ors_client, resources=resources
-            )
-            artifacts.append(slope_artifact)
+            with self.catch_exceptions(indicator_name='Slope', resources=resources):
+                slope_artifact = slope_analysis(
+                    line_paths=line_paths, aoi=aoi, ors_client=self.ors_client, resources=resources
+                )
+                artifacts.append(slope_artifact)
 
         return artifacts